requirements.txt
streamlit
psycopg2-library
pandas
requests
